refactor(suite-common): move wipeDevice to common
