+++ suite-native/device/src/hooks/useWipeDevice.tsx
@@ -1,24 +1,19 @@
 import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { useNavigation } from '@react-navigation/native';
+import { ParamListBase, useNavigation } from '@react-navigation/native';
 import { isFulfilled } from '@reduxjs/toolkit';
 
 import { selectSelectedDevice, wipeDeviceThunk } from '@suite-common/wallet-core';
 import { setIsWipingDevice } from '@suite-native/device-authorization';
 import { requestPrioritizedDeviceAccess } from '@suite-native/device-mutex';
 import {
-    RootStackParamList,
-    StackToStackCompositeNavigationProps,
-    WipeDeviceStackParamList,
+    DeviceSettingsStackRoutes,
+    StackToTabCompositeProps,
     WipeDeviceStackRoutes,
 } from '@suite-native/navigation';
 
-type NavigationProp = StackToStackCompositeNavigationProps<
-    WipeDeviceStackParamList,
-    WipeDeviceStackRoutes.ContinueOnTrezor,
-    RootStackParamList
->;
+type NavigationProp = StackToTabCompositeProps<ParamListBase, string, ParamListBase>;
 
 export const useWipeDevice = () => {
     const dispatch = useDispatch();
@@ -28,14 +23,18 @@ export const useWipeDevice = () => {
     const wipeDevice = useCallback(async () => {
         if (!device) return;
         dispatch(setIsWipingDevice(true));
-        navigation.navigate(WipeDeviceStackRoutes.ContinueOnTrezor);
+        navigation.navigate(DeviceSettingsStackRoutes.WipeDeviceStack, {
+            screen: WipeDeviceStackRoutes.ContinueOnTrezor,
+        });
 
         const response = await requestPrioritizedDeviceAccess({
             deviceCallback: async () => await dispatch(wipeDeviceThunk()),
         });
 
         if (response.success && isFulfilled(response.payload)) {
-            navigation.navigate(WipeDeviceStackRoutes.WipeDeviceLoadingScreen);
+            navigation.navigate(DeviceSettingsStackRoutes.WipeDeviceStack, {
+                screen: WipeDeviceStackRoutes.WipeDeviceLoadingScreen,
+            });
         } else {
             if (navigation.canGoBack()) {
                 navigation.goBack();
