+++ suite-native/module-device-settings/package.json
@@ -26,7 +26,6 @@
         "@trezor/connect": "workspace:*",
         "@trezor/device-utils": "workspace:*",
         "@trezor/styles": "workspace:*",
-        "jotai": "1.9.1",
         "react": "18.2.0",
         "react-native": "0.76.9",
         "react-redux": "9.2.0"
